contact page started on + partnerships component created
More updates, a contact page is being worked on where the states are
being worked on but now need to consider whether we want to email the
society from the forms or possibly create a sticky note type element
where it will remain on the site but will have to research it.
Partnership component has been created using AIAA and Bluesat logo +
text but can be edited more.